q4,q6,q7 checked and updated
q4 was actually correct, it should display all the books but the ones that don't belong to Science Fiction nor Fantasy.
for q6 at first I thought maybe I should consider genre as a subcategory but then I figured that it's better to just display whatever it is a main or individual, not subcategory.